+++ rust/dbn-cli/src/lib.rs
@@ -1,6 +1,7 @@
 use std::{
     fs::File,
     io::{self, BufWriter},
+    num::NonZeroU64,
     path::PathBuf,
 };
 
@@ -89,6 +90,14 @@ pub struct Args {
          help ="Make the CSV or JSON output easier to read by converting timestamps to ISO 8601 and prices to decimals"
     )]
     pub should_pretty_print: bool,
+    #[clap(
+        short = 'l',
+        long = "limit",
+        value_name = "NUM",
+        conflicts_with = "should_output_metadata",
+        help = "Limit the number of records in the output to the specified number"
+    )]
+    pub limit: Option<NonZeroU64>,
 }
 
 impl Args {